Refactor slixer commands util folder structure
